Remove grouped des performance results
I've just noticed that the grouped deserialization throughput benchmark was unfair for Jena, because it used a Jena parser API that was spawning a separate parser thread for each graph/dataset parsed. This added up to horrendous overhead, which explains the weird performance gap. This API in Jena is very weird and I'm really surprised it manages threads on its own. Why would it do that? In any case, Jelly is still faster...