RDF spec: fix inconsistency in name lookup sizing (#35)
* RDF spec: fix inconsistency in name lookup sizing

The spec said that the name lookup could be 0-sized, which is not really possible in practice, as in this case, it's impossible to encode any IRIs. Any well-formed RDF requires the use of IRIs (in the predicate position), so that makes no sense.

There was also a bit of wrong text about how the name lookup references in RdfIri should be interpreted, I fixed that as well.

* Update docs/specification/serialization.md

Co-authored-by: Anastasiya Danilenka <41057639+adanilenka@users.noreply.github.com>

---------

Co-authored-by: Anastasiya Danilenka <41057639+adanilenka@users.noreply.github.com>